Started work on discreet private messaging
The goal now is to make the private messaging more elegant. So far,
private messages appear in a separate window with a dynamically
generated id. This is about 25% of the way towards completion.

The main issue is to create a new tab / window for every private
messaging session, so that users can skip back and forth between
multiple conversations.