Tabs Added (Somewhat)
Spent 3 hours going back and forth trying to work out how to dynamically
add tabs. So far tabs are being added with notifications, multiple tabs
with the right names assigned to them. However, there is no way to
remove a tab at the moment and content is still all being beamed to a
single div instead of being assigned to discreet tabs (which will be the
next step).

Once tabs are sorted project 'snooper.co' will go live on a server to
begin version 0.1 testing ;)