Added Dynamic Divs
Added dynamic divs so that a private message div exists for every user,
and if the user is self then the display for that div is none. Messages
are being appended to the new divs based upon who the message is from,
so the next step is to make it so that display is toggled depending on
which user is selected. After that, the trick will be to make clicking
on divs the toggle event.